--- HealingPotion.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Engine
-{
-    public class HealingPotion : Item
-    {
-        public int AmountToHeal { get; set; }
-    }
-}
