--- Item.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Engine
-{
-    public class Item
-    {
-        public int ID { get; set; }
-        public string Name { get; set; }
-        public string NamePlural { get; set; }
-    }
-}
-
