--- Monster.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Engine
-{
-    public class Monster : LivingCreature
-    {
-        public int ID { get; set; }
-        public string Name { get; set; }
-        public int MaximumDamage { get; set; }
-        public int RewardExperiencePoints { get; set; }
-        public int RewardGold { get; set; }
-    }
-}
