--- SuperAdventure.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.31025.194
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SuperAdventure", "SuperAdventure.csproj", "{94D832B8-9552-4379-B137-38EE1F5B2BD7}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Engine", "..\Engine\Engine.csproj", "{588437C5-18A5-4647-A57A-501776D49BF4}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{94D832B8-9552-4379-B137-38EE1F5B2BD7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{94D832B8-9552-4379-B137-38EE1F5B2BD7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{94D832B8-9552-4379-B137-38EE1F5B2BD7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{94D832B8-9552-4379-B137-38EE1F5B2BD7}.Release|Any CPU.Build.0 = Release|Any CPU
-		{588437C5-18A5-4647-A57A-501776D49BF4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{588437C5-18A5-4647-A57A-501776D49BF4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{588437C5-18A5-4647-A57A-501776D49BF4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{588437C5-18A5-4647-A57A-501776D49BF4}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {46ADC684-BD4B-4769-B546-BAE0EF2F321A}
-	EndGlobalSection
-EndGlobal
