+++ SuperAdventure/Engine/HealingPotion.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Engine
+{
+    public class HealingPotion : Item
+    {
+        public int AmountToHeal { get; set; }
+    }
+}
