+++ SuperAdventure/Engine/LivingCreature.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace Engine
+{
+    public class LivingCreature
+    {
+        public int CurrentHitPoints { get; set; }
+        public int MaximumHitPoints { get; set; }
+    }
+}
