+++ SuperAdventure/Engine/Weapon.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Engine
+{
+    public class Weapon : Item
+    {
+        public int MinimumDamage { get; set; }
+        public int MaximumDamage { get; set; }
+    }
+}
