+++ SuperAdventure/SuperAdventure.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31005.135
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Engine", "Engine\Engine.csproj", "{EFCB9D64-5B06-4A0A-974A-361FEDAF7DBB}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SuperAdventure", "SuperAdventure\SuperAdventure.csproj", "{753166F5-E2E3-4485-A47D-A37292C8CCC2}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{EFCB9D64-5B06-4A0A-974A-361FEDAF7DBB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EFCB9D64-5B06-4A0A-974A-361FEDAF7DBB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EFCB9D64-5B06-4A0A-974A-361FEDAF7DBB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EFCB9D64-5B06-4A0A-974A-361FEDAF7DBB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{753166F5-E2E3-4485-A47D-A37292C8CCC2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{753166F5-E2E3-4485-A47D-A37292C8CCC2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{753166F5-E2E3-4485-A47D-A37292C8CCC2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{753166F5-E2E3-4485-A47D-A37292C8CCC2}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {D20C7C3E-FFF1-49EE-9E38-EC93F28BAFF7}
+	EndGlobalSection
+EndGlobal
